+++ sys/fs/unicode.h
@@ -1,4 +1,4 @@
-/* $NetBSD: unicode.h,v 1.5 2007/02/26 12:33:08 drochner Exp $ */
+/* $NetBSD: unicode.h,v 1.7 2014/04/06 19:25:22 jakllsch Exp $ */
 
 /*-
  * Copyright (c) 2001, 2004 The NetBSD Foundation, Inc.
@@ -12,13 +12,6 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *        This product includes software developed by the NetBSD
- *        Foundation, Inc. and its contributors.
- * 4. Neither the name of The NetBSD Foundation nor the names of its
- *    contributors may be used to endorse or promote products derived
- *    from this software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE NETBSD FOUNDATION, INC. AND CONTRIBUTORS
  * ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
@@ -83,7 +76,7 @@ static int wput_utf8(char *, size_t, u_int16_t) __unused;
 static u_int16_t
 wget_utf8(const char **str, size_t *sz)
 {
-	int c;
+	size_t c;
 	u_int16_t rune = 0;
 	const char *s = *str;
 	static const int _utf_count[16] = {
